Extract and classify git commits as contributions

+++ src/main/kotlin/no/masterthesis/domain/changecontribution/ChangeContribution.kt
@@ -36,11 +36,11 @@ data class ChangeContribution(
    * The number of lines added for the specified commit and filePath.
    * It should adjust for renames/file moves.
    * */
-  val linesAdded: Long,
+  val linesAdded: Int,
   /**
    * Similar to [linesAdded] but for lines removed.
    * */
-  val linesRemoved: Long,
+  val linesRemoved: Int,
 
   /**
    * Flag used to tell what the previous filePath was,
